Enemy AI working, working on squad behavior now

+++ Assets/Scripts/CS_Tile.cs
@@ -7,7 +7,7 @@ public class Tile
     public Tile parent;
     public MB_Entity entity = null;
     public Vector2Int position;
-    public int costFromOrigin = 0;
+    public float costFromOrigin = 0;
     public float costToDestination = 0;
     public int terrainCost = 0;
     public float TotalCost { get { return costFromOrigin + costToDestination + terrainCost; } }
